refactor: dedupe step5/step5B and drop dead internal utilities
Extract the byte-identical step5/step5B augment helpers into
src/core/shared.ts and import them from all six algorithm variants,
removing five redundant copies. The per-cell numeric kernels stay split
per number/bigint/finite domain (deliberate performance separation), so
only the pointer-shuffle helpers are shared.

Remove unused internal utilities pad, padHeight, padWidth, rot90,
rotNeg90, toString (utils/matrix.ts) and isNumber (utils/is.ts) plus
their tests. All are internal (not part of the published index.ts
surface), so no consumer-facing behavior change. Net ~725 fewer lines.
