+++ src/core/numFiniteMunkres.ts
@@ -24,6 +24,7 @@ import { getMin } from "../utils/arrayLike.ts";
 import { partitionByMin } from "../utils/mutableArrayLike.ts";
 
 import { step4B } from "./numFiniteMunkresB.ts";
+import { step5 } from "./shared.ts";
 
 export function exec(matrix: MatrixLike<number>): Matching<number> {
   // Get dimensions
@@ -196,36 +197,6 @@ export function step4(
     // Update matching
     step5(slack[N - 1], slackY, starsX, starsY);
   }
-}
-
-/**
- * Augments the current matching.
- *
- * This step effectively increases the number of matches (stars)
- * by 1, bringing the algorithm closer to an optimal assignment.
- *
- * Augmentation is performed by flipping matched and unmatched edges along
- * an augmenting path, starting from an unmatched node / edge and
- * continuing until no matched edge can be found.
- *
- * @param x - The starting node's column.
- * @param primeX - An array mapping primed columns to rows.
- * @param starsX - An array mapping star columns to row. Modified in place.
- * @param starsY - An array mapping star rows to columns. Modified in place.
- */
-export function step5(
-  x: number,
-  primeX: ArrayLike<number>,
-  starsX: MutableArrayLike<number>,
-  starsY: MutableArrayLike<number>,
-): void {
-  do {
-    const y = primeX[x];
-    const sx = starsY[y];
-    starsX[x] = y;
-    starsY[y] = x;
-    x = sx;
-  } while (x !== -1);
 }
 
 /**
