+++ src/core/shared.ts
@@ -1,0 +1,61 @@
+import type { MutableArrayLike } from "../types/mutableArrayLike.ts";
+
+/**
+ * Augments the current matching.
+ *
+ * This step effectively increases the number of matches (stars)
+ * by 1, bringing the algorithm closer to an optimal assignment.
+ *
+ * Augmentation is performed by flipping matched and unmatched edges along
+ * an augmenting path, starting from an unmatched node / edge and
+ * continuing until no matched edge can be found.
+ *
+ * @param x - The starting node's column.
+ * @param primeX - An array mapping primed columns to rows.
+ * @param starsX - An array mapping star columns to row. Modified in place.
+ * @param starsY - An array mapping star rows to columns. Modified in place.
+ */
+export function step5(
+  x: number,
+  primeX: ArrayLike<number>,
+  starsX: MutableArrayLike<number>,
+  starsY: MutableArrayLike<number>,
+): void {
+  do {
+    const y = primeX[x];
+    const sx = starsY[y];
+    starsX[x] = y;
+    starsY[y] = x;
+    x = sx;
+  } while (x !== -1);
+}
+
+/**
+ * Augments the current matching.
+ *
+ * This step effectively increases the number of matches (stars)
+ * by 1, bringing the algorithm closer to an optimal assignment.
+ *
+ * Augmentation is performed by flipping matched and unmatched edges along
+ * an augmenting path, starting from an unmatched node / edge and
+ * continuing until no matched edge can be found.
+ *
+ * @param y - The starting node's row.
+ * @param primeY - An array mapping primed rows to columns.
+ * @param starsX - An array mapping star columns to row. Modified in place.
+ * @param starsY - An array mapping star rows to columns. Modified in place.
+ */
+export function step5B(
+  y: number,
+  primeY: ArrayLike<number>,
+  starsX: MutableArrayLike<number>,
+  starsY: MutableArrayLike<number>,
+): void {
+  do {
+    const x = primeY[y];
+    const sy = starsX[x];
+    starsX[x] = y;
+    starsY[y] = x;
+    y = sy;
+  } while (y !== -1);
+}
