+++ src/utils/is.ts
@@ -14,20 +14,3 @@
 export function isBigInt(value: unknown): value is bigint {
   return typeof value === "bigint";
 }
-
-/**
- * Checks if the given value is of type `number`.
- *
- * @param value - The value to check.
- *
- * @returns `true` if the value is of type `number`, `false` otherwise.
- *
- * @example
- * console.log(isNumber(10)); // true
- *
- * @example
- * console.log(isNumber(10n)); // false
- */
-export function isNumber(value: unknown): value is number {
-  return typeof value === "number";
-}
