+++ src/utils/matrix.test.ts
@@ -12,12 +12,6 @@ import {
   invert,
   map,
   negate,
-  pad,
-  padHeight,
-  padWidth,
-  rot90,
-  rotNeg90,
-  toString,
   transpose,
 } from "./matrix.ts";
 import { getMax } from "./matrixLike.ts";
@@ -607,385 +601,6 @@ describe(`${negate.name}()`, () => {
     ];
     negate(mat);
     expect(mat).toEqual(expected);
-  });
-});
-
-describe(`${pad.name}()`, () => {
-  test("correctly pads a matrix to the specified width and height", () => {
-    const matrix = [
-      [1, 2],
-      [3, 4],
-    ];
-    pad(matrix, 3, 4, 0);
-    const expectedMatrix = [
-      [1, 2, 0, 0],
-      [3, 4, 0, 0],
-      [0, 0, 0, 0],
-    ];
-    expect(matrix).toEqual(expectedMatrix);
-  });
-
-  test("pads width only when height already meets the specified value", () => {
-    const matrix = [
-      [1, 2, 3],
-      [4, 5, 6],
-    ];
-    pad(matrix, 2, 5, 0);
-    const expectedMatrix = [
-      [1, 2, 3, 0, 0],
-      [4, 5, 6, 0, 0],
-    ];
-    expect(matrix).toEqual(expectedMatrix);
-  });
-
-  test("pads height only when width already meets the specified value", () => {
-    const matrix = [
-      [1, 2, 3, 4],
-      [5, 6, 7, 8],
-    ];
-    pad(matrix, 4, 4, 0);
-    const expectedMatrix = [
-      [1, 2, 3, 4],
-      [5, 6, 7, 8],
-      [0, 0, 0, 0],
-      [0, 0, 0, 0],
-    ];
-    expect(matrix).toEqual(expectedMatrix);
-  });
-
-  test("does not modify the matrix if it already meets the specified dimensions", () => {
-    const matrix = [
-      [1, 2, 3],
-      [4, 5, 6],
-      [7, 8, 9],
-    ];
-    pad(matrix, 3, 3, 0);
-    const expectedMatrix = [
-      [1, 2, 3],
-      [4, 5, 6],
-      [7, 8, 9],
-    ];
-    expect(matrix).toEqual(expectedMatrix);
-  });
-
-  test("handles an empty matrix by padding it to the specified dimensions", () => {
-    const matrix: Matrix<string> = [];
-    pad(matrix, 2, 3, "c"); // Pad an empty matrix to 2x3 with 'c'
-    const expectedMatrix = [
-      ["c", "c", "c"],
-      ["c", "c", "c"],
-    ];
-    expect(matrix).toEqual(expectedMatrix);
-  });
-});
-
-describe(`${padHeight.name}()`, () => {
-  test("adds rows to meet the specified height with fill value", () => {
-    const matrix = [
-      [1, 2],
-      [3, 4],
-    ];
-    padHeight(matrix, 4, 0);
-    const expectedMatrix = [
-      [1, 2],
-      [3, 4],
-      [0, 0],
-      [0, 0],
-    ];
-    expect(matrix).toEqual(expectedMatrix);
-  });
-
-  test("does not modify matrix if it already meets the specified height", () => {
-    const originalMatrix = [
-      [1, 2],
-      [3, 4],
-      [5, 6],
-    ];
-    padHeight(originalMatrix, 3, 0);
-    expect(originalMatrix).toEqual([
-      [1, 2],
-      [3, 4],
-      [5, 6],
-    ]);
-  });
-
-  test("does not modify matrix if it already exceeds the specified height", () => {
-    const originalMatrix = [
-      [1, 2],
-      [3, 4],
-      [5, 6],
-    ];
-    padHeight(originalMatrix, 1, 0);
-    expect(originalMatrix).toEqual([
-      [1, 2],
-      [3, 4],
-      [5, 6],
-    ]);
-  });
-
-  test("handles empty matrix by filling up to the specified height", () => {
-    const matrix: Matrix<string> = [];
-    padHeight(matrix, 2, "empty");
-    const expectedMatrix = [[], []];
-    expect(matrix).toEqual(expectedMatrix);
-  });
-});
-
-describe(`${padWidth.name}()`, () => {
-  test("adds columns to meet the specified width with fill value", () => {
-    const matrix = [
-      [1, 2],
-      [3, 4],
-    ];
-    padWidth(matrix, 4, 0);
-    const expectedMatrix = [
-      [1, 2, 0, 0],
-      [3, 4, 0, 0],
-    ];
-    expect(matrix).toEqual(expectedMatrix);
-  });
-
-  test("does not modify matrix if it already meets the specified width", () => {
-    const originalMatrix = [
-      [1, 2, 3],
-      [4, 5, 6],
-    ];
-    padWidth(originalMatrix, 3, 0);
-    expect(originalMatrix).toEqual([
-      [1, 2, 3],
-      [4, 5, 6],
-    ]);
-  });
-
-  test("does not modify matrix if it already exceeds the specified width", () => {
-    const originalMatrix = [
-      [1, 2, 3],
-      [4, 5, 6],
-    ];
-    padWidth(originalMatrix, 1, 0);
-    expect(originalMatrix).toEqual([
-      [1, 2, 3],
-      [4, 5, 6],
-    ]);
-  });
-
-  test("does not modify an empty matrix", () => {
-    const matrix: Matrix<string> = [];
-    padWidth(matrix, 2, "empty");
-    const expectedMatrix: Matrix<string> = [];
-    expect(matrix).toEqual(expectedMatrix);
-  });
-});
-
-describe(`${rot90.name}()`, () => {
-  test("handles an empty matrix", () => {
-    const mat: Matrix<number> = [];
-    rot90(mat);
-    expect(mat).toEqual([]);
-  });
-
-  test("handles a 1x1 matrix", () => {
-    const mat = [[2]];
-    rot90(mat);
-    expect(mat).toEqual([[2]]);
-  });
-
-  test("rotates a 2x2 matrix by 90 degrees clockwise", () => {
-    const mat = [
-      [1, 2],
-      [3, 4],
-    ];
-    rot90(mat);
-    expect(mat).toEqual([
-      [3, 1],
-      [4, 2],
-    ]);
-  });
-
-  test("rotates a 2x3 matrix by 90 degrees clockwise", () => {
-    const mat = [
-      [1, 2, 3],
-      [4, 5, 6],
-    ];
-    rot90(mat);
-    expect(mat).toEqual([
-      [4, 1],
-      [5, 2],
-      [6, 3],
-    ]);
-  });
-
-  test("rotates a 3x2 matrix by 90 degrees clockwise", () => {
-    const mat = [
-      [1, 2],
-      [3, 4],
-      [5, 6],
-    ];
-    rot90(mat);
-    expect(mat).toEqual([
-      [5, 3, 1],
-      [6, 4, 2],
-    ]);
-  });
-
-  test("rotates a 3x3 matrix by 90 degrees clockwise", () => {
-    const mat = [
-      [1, 2, 3],
-      [4, 5, 6],
-      [7, 8, 9],
-    ];
-    rot90(mat);
-    expect(mat).toEqual([
-      [7, 4, 1],
-      [8, 5, 2],
-      [9, 6, 3],
-    ]);
-  });
-
-  test("correctly rotates a 4x1 matrix by 90 degrees clockwise", () => {
-    const mat = [[1], [2], [3], [4]];
-    rot90(mat);
-    expect(mat).toEqual([[4, 3, 2, 1]]);
-  });
-
-  test("correctly rotates a 1x4 matrix by 90 degrees clockwise", () => {
-    const mat = [[1, 2, 3, 4]];
-    rot90(mat);
-    expect(mat).toEqual([[1], [2], [3], [4]]);
-  });
-});
-
-describe(`${rotNeg90.name}()`, () => {
-  test("handles an empty matrix", () => {
-    const mat: Matrix<number> = [];
-    rotNeg90(mat);
-    expect(mat).toEqual([]);
-  });
-
-  test("handles a 1x1 matrix", () => {
-    const mat = [[2]];
-    rotNeg90(mat);
-    expect(mat).toEqual([[2]]);
-  });
-
-  test("rotates a 2x2 matrix by 90 degrees counterclockwise", () => {
-    const mat = [
-      [1, 2],
-      [3, 4],
-    ];
-    rotNeg90(mat);
-    expect(mat).toEqual([
-      [2, 4],
-      [1, 3],
-    ]);
-  });
-
-  test("rotates a 2x3 matrix by 90 degrees counterclockwise", () => {
-    const mat = [
-      [1, 2, 3],
-      [4, 5, 6],
-    ];
-    rotNeg90(mat);
-    expect(mat).toEqual([
-      [3, 6],
-      [2, 5],
-      [1, 4],
-    ]);
-  });
-
-  test("rotates a 3x2 matrix by 90 degrees counterclockwise", () => {
-    const mat = [
-      [1, 2],
-      [3, 4],
-      [5, 6],
-    ];
-    rotNeg90(mat);
-    expect(mat).toEqual([
-      [2, 4, 6],
-      [1, 3, 5],
-    ]);
-  });
-
-  test("rotates a 3x3 matrix by 90 degrees counterclockwise", () => {
-    const mat = [
-      [1, 2, 3],
-      [4, 5, 6],
-      [7, 8, 9],
-    ];
-    rotNeg90(mat);
-    expect(mat).toEqual([
-      [3, 6, 9],
-      [2, 5, 8],
-      [1, 4, 7],
-    ]);
-  });
-
-  test("rotates a 1x4 matrix by 90 degrees counterclockwise", () => {
-    const mat = [[1, 2, 3, 4]];
-    rotNeg90(mat);
-    expect(mat).toEqual([[4], [3], [2], [1]]);
-  });
-
-  test("rotates a 4x1 matrix by 90 degrees counterclockwise", () => {
-    const mat = [[1], [2], [3], [4]];
-    rotNeg90(mat);
-    expect(mat).toEqual([[1, 2, 3, 4]]);
-  });
-});
-
-describe(`${toString.name}()`, () => {
-  test("handles an empty matrix", () => {
-    const mat: Matrix<number> = [];
-    const expectedString = "";
-    expect(toString(mat)).toEqual(expectedString);
-  });
-
-  test("handles a 1x1 matrix", () => {
-    const mat = [[2]];
-    const expectedString = "[2]";
-    expect(toString(mat)).toEqual(expectedString);
-  });
-
-  test("processes a single row matrix", () => {
-    const mat = [[1, 2, 3]];
-    const expectedString = "[1, 2, 3]";
-    expect(toString(mat)).toEqual(expectedString);
-  });
-
-  test("processes a single column matrix", () => {
-    const mat = [[1], [2], [3]];
-    const expectedString = "[1],\n" + "[2],\n" + "[3]";
-    expect(toString(mat)).toEqual(expectedString);
-  });
-
-  test("converts a numeric matrix to a string representation", () => {
-    const mat = [
-      [1, 2, 3],
-      [4, 5, 6],
-    ];
-    const expectedString = "[1, 2, 3],\n[4, 5, 6]";
-    expect(toString(mat)).toEqual(expectedString);
-  });
-
-  test("converts a matrix with mixed types to a string representation", () => {
-    const mat = [
-      ["a", 2, true],
-      [null, undefined, 5],
-    ];
-    const expectedString =
-      "[   a,         2, true],\n" + "[null, undefined,    5]";
-    expect(toString(mat)).toEqual(expectedString);
-  });
-
-  test("aligns values properly in columns when converting to string", () => {
-    const mat = [
-      [1, 100, 1000],
-      [20, 200, 20000],
-      [300, 3000, 300],
-    ];
-    const expectedString =
-      "[  1,  100,  1000],\n" + "[ 20,  200, 20000],\n" + "[300, 3000,   300]";
-    expect(toString(mat)).toEqual(expectedString);
   });
 });
