+++ src/utils/matrix.ts
@@ -300,203 +300,6 @@ export function negate<T extends number | bigint>(matrix: Matrix<T>): void {
 }
 
 /**
- * Pads a matrix to a specified size with a given fill value.
- *
- * The padding is applied from the ends (right) of each row and
- * the ends (bottom) of each column. If a dimension is already
- * at or above the desired value, no change is made to it.
- *
- * @param matrix - The matrix to pad. Modified in place.
- * @param height - The desired number of rows in the matrix.
- * @param width - The desired number of columns in the matrix.
- * @param fillValue - The value used for padding.
- */
-export function pad<T>(
-  matrix: Matrix<T>,
-  height: number,
-  width: number,
-  fillValue: T,
-): void {
-  padHeight(matrix, height, fillValue);
-  padWidth(matrix, width, fillValue);
-}
-
-/**
- * Pads the height (number of rows) of a matrix with a given fill value.
- *
- * Rows are added to the end (bottom) of the matrix until its height reaches
- * `height`, with each new row filled with `fillValue`. If the matrix is
- * already at or above `height`, no change is made.
- *
- * @param matrix - The matrix to pad. Modified in place.
- * @param height - The desired number of rows in the matrix.
- * @param fillValue - The value to use for filling new rows.
- */
-export function padHeight<T>(
-  matrix: Matrix<T>,
-  height: number,
-  fillValue: T,
-): void {
-  const Y = matrix.length;
-  if (Y >= height) {
-    return;
-  }
-
-  matrix.length = height;
-  const X = matrix[0]?.length ?? 0;
-  for (let y = Y; y < height; ++y) {
-    matrix[y] = new Array<T>(X).fill(fillValue);
-  }
-}
-
-/**
- * Pads the width (number of columns) of a matrix with a given fill value.
- *
- * Columns are added to the right of the matrix until its width reaches
- * `width`, with each new column filled with `fillValue`. If the matrix is
- * already at or above `width`, no change is made.
- *
- * @param matrix - The matrix to pad. Modified in place.
- * @param width - The desired number of columns in the matrix.
- * @param fillValue - The value to use for filling new columns.
- */
-export function padWidth<T>(
-  matrix: Matrix<T>,
-  width: number,
-  fillValue: T,
-): void {
-  const X = matrix[0]?.length ?? 0;
-  if (X >= width) {
-    return;
-  }
-
-  const Y = matrix.length;
-  for (let y = 0; y < Y; ++y) {
-    matrix[y].length = width;
-    matrix[y].fill(fillValue, X, width);
-  }
-}
-
-/**
- * Rotates a matrix by 90 degrees clockwise.
- *
- * @param matrix - The matrix to be rotated. Modified in place.
- *
- * @example
- * const matrix = [
- *   [1, 2],
- *   [3, 4]
- * ];
- *
- * rot90(matrix);
- * // matrix is now:
- * // [
- * //   [3, 1],
- * //   [4, 2]
- * // ]
- *
- * @example
- * const matrix = [
- *   [1, 2, 3],
- *   [4, 5, 6]
- * ];
- *
- * rot90(matrix);
- * // matrix is now:
- * // [
- * //   [4, 1],
- * //   [5, 2],
- * //   [6, 3]
- * // ]
- */
-export function rot90<T>(matrix: Matrix<T>): void {
-  flipV(matrix);
-  transpose(matrix);
-}
-
-/**
- * Rotates a matrix by 90 degrees counterclockwise.
- *
- * @param matrix - The matrix to be rotated. Modified in place.
- *
- * @example
- * const matrix = [
- *   [1, 2],
- *   [3, 4]
- * ];
- *
- * rot90(matrix);
- * // matrix is now:
- * // [
- * //   [2, 4],
- * //   [1, 3]
- * // ]
- *
- * @example
- * const matrix = [
- *   [1, 2, 3],
- *   [4, 5, 6]
- * ];
- *
- * rot90(matrix);
- * // matrix is now:
- * // [
- * //   [3, 6],
- * //   [2, 5],
- * //   [1, 4]
- * // ]
- */
-export function rotNeg90<T>(matrix: Matrix<T>): void {
-  transpose(matrix);
-  flipV(matrix);
-}
-/**
- * Generates a string representation of a matrix.
- *
- * @param matrix - The matrix.
- * @param callbackFn - (Optional) A callback function to convert each element
- * to a string. Defaults to using each elements `toString` method.
- *
- * @returns A string representation of the matrix.
- */
-export function toString<T>(
-  matrix: MatrixLike<T>,
-  callbackFn: (
-    value: T,
-    row: number,
-    col: number,
-    mat: typeof matrix,
-  ) => string = (v) => `${v}`,
-): string {
-  const strs: Matrix<string> = map(matrix, callbackFn);
-  const Y = strs.length;
-  const X = strs[0]?.length ?? 0;
-
-  // For each column
-  for (let x = 0; x < X; ++x) {
-    // Get width
-    let width = 0;
-    for (let y = 0; y < Y; ++y) {
-      width = Math.max(width, strs[y][x].length);
-    }
-
-    // Adjust width
-    for (let y = 0; y < Y; ++y) {
-      strs[y][x] = strs[y][x].padStart(width, " ");
-    }
-  }
-
-  // Create output
-  const buf: string[] = new Array(Y);
-  for (let y = 0; y < Y; ++y) {
-    buf[y] = `[${strs[y].join(", ")}]`;
-  }
-
-  // Return output
-  return buf.join(",\n");
-}
-
-/**
  * Transpose a given matrix, switching its rows and columns.
  *
  * In the transposed matrix, the value originally at position [y][x]
